+++ flutter_application_practice/lib/screens/favorites_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+import '../widgets/drawer_menu.dart';
+import '../widgets/favorites_page.dart';
+
+class FavoritesScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return DrawerMenu(screen: FavoritesPage(), selectedIndex: 1);
+  }
+}
