+++ flutter_application_practice/lib/screens/home_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import '../widgets/drawer_menu.dart';
+import '../widgets/generator_page.dart';
+
+class MyHomeScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return DrawerMenu(
+      screen: GeneratorPage(),
+      selectedIndex: 0,
+    );
+  }
+}
