+++ flutter_application_practice/lib/widgets/big_card.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:english_words/english_words.dart';
+
+class BigCard extends StatelessWidget {
+  const BigCard({
+    super.key,
+    required this.pair,
+  });
+
+  final WordPair pair;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final style = theme.textTheme.displayMedium!
+        .copyWith(color: theme.colorScheme.onPrimary, letterSpacing: 4);
+    return Card(
+      color: theme.colorScheme.primary,
+      elevation: 15,
+      child: Padding(
+        padding: const EdgeInsets.all(20),
+        child: Text(
+          pair.asCamelCase,
+          style: style,
+          semanticsLabel: "${pair.first} ${pair.second}",
+        ),
+      ),
+    );
+  }
+}
