+++ flutter_application_practice/pubspec.yaml
@@ -26,3 +26,5 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
+
+assets:
